check-list/requirements.txt
psycopg2-binary
# mariadb

